knotty.py: Extend knotty to support "real time" log tail for tasks
This patch provides the capability to knotty to have a higher log
level where the individual tasks logs can be multiplexed to stdio.
This is highly desirable for several different use cases.

1) You can visually spot "hot spots" in building parallel tasks

2) The relative order an sequence of events can be determined in
   difficult to diagnose inter-related library problems.

3) It provides a nice way to see what all is actually going on and
   getting logged by various tasks such as a kernel compile where you
   might want to actually look at the output in a build over build
   scenario.

4) There are most certainly other aspects where this is useful.

The primary control is to set the BB_RT_LOGLEVEL = "2" in your
local.conf.  Additional capabilities will follow in successive
patches.  When parallel tasks are enabled, the logs will be
multiplexed and prefixed with the PID of the task.  In the
non-parallel case there is no PID prefix as everything is linear.

(LOCAL REV: NOT UPSTREAM) -- Sent to bitbake-devel on 20120917

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>

TaskStarted __init__ is no long necessary since the base class
now includes the logfile.

Signed-off-by: Jeff Polk <jeff.polk@windriver.com>
[MAA: context updates after upstream moved to py logging]
Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>