msg.py, knotty.py: Allow dynamic toggle of the debug log level
If stdin is a controlling tty, make it possible to have
log levels 1-4 where you can press a key to dynamically
change the log level, while bitbake is running to see
what is going on.

1 = default log level
2 = multiple log tail of bitbake tasks
3 = bit bake debug messages
4 = bitbake debug messages + multiple lot tail of bitbake tasks

In order to make this possible a slight modification to
the msg logfilter was needed to make it possible to get
an object so as to change the filter log level.

(LOCAL REV: NOT UPSTREAM) -- Sent to bitbake-devel on 20120917

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>