fetch2/git.py: fix the filename too long error
CQID: WIND00440943

When we use the PREMIRROR/MIRROR from the local disk, and if it is in a
deep directory, for example, when len(path) > 255 (The NAME_MAX), we
will get the "filename too long error".

This is becuase we use ud.path.replace('/', '.') to change the path to
the filename. We seldom see such a long url, but it is easy to produce
it on the local machine.

Another problem is that:

file:///local/path

will be converted into .local.path which is a hidden file by default.

Don't convert the name when ud.proto == "file" will fix the problem.

(LOCAL REV: NOT UPSTREAM) -- Sent to bitbake-devel on 20131108

Signed-off-by: Robert Yang <liezhi.yang@windriver.com>