Toaster: add progress bar for cloning
Issue: LIN9-4460

With the large number of layers in the default project, the
build appear to hang while those layers are checked out
(especially for the kernel).
This patch adds a clone progress bar that is visible before
the parsing progress appears.

Also, the Wind River specific database migration file is
moved to "0100_wrtemplates.py" so that upstream migrations
can stay in sequence.

(LOCAL REV: NOT UPSTREAM) -- Sent to toaster@yoctoproject.org on 20170621

Signed-off-by: David Reyna <David.Reyna@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>