LIN9-4351 Toaster OperationalError: too many SQL variables
Toaster: too many SQL variables for package set

Issue: LIN9-4351

If you build a project with a large package set, you will get a crash
in "views.py" when the dashboard attempts to fetch the package set to
calculate the package count and size. This is a sqlite limitation, and
it fails with as few as 1220 packages.

(LOCAL REV: NOT UPSTREAM) -- Sent to toaster@yoctoproject.org on 20170623

Signed-off-by: David Reyna <David.Reyna@windriver.com>