Toaster: custom image updates and original creation
Issue: LIN9-4597 LIN9-4481

When Toaster removes a package from a custom image, it must
also always remove the advised reverse-dependent recipes.
Similarly, when adding a package it must always add its advised
the packages it depends on.

Toaster normaly waited until a new custom image was built before
creating the custom layer and the recipe. However, a different build
can fail because the recipe has already been added to the project, so
the image's default recipe must be created when the image is created.

(LOCAL REV: NOT UPSTREAM) -- Sent to toaster@yoctoproject.org on 20170808

Signed-off-by: David Reyna <David.Reyna@windriver.com>