fetch2/gitsm.py: Disable branch checking on submodules
Submodules by definition refer to a specific commit, not branch.  If we don't
ignore the branch, then any commits on a submodule on a branch different then
the original module will trigger a failure that the commit is not on the
branch.

(LOCAL REV: NOT UPSTREAM) -- Sent to bitbake-devel on 20181031

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>