knotty.py: use the already fetched BB_RT_LOGLEVEL
The "getVariable" of BB_RT_LOGLEVEL later in the main() function
potentially causes issues with bitbake. We already take the time to
fetch this varilable at the top of main() along with other existing
"getVariable" queries, so use the variable as we should have in the
first place, and avoid the fetch later in main().

Signed-off-by: Mark Asselstine <mark.asselstine@windriver.com>